# server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "adhocly-server",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local adhocly API"
    }
  ],
  "paths": {
    "/api/boards": {
      "get": {
        "tags": [
          "kanban"
        ],
        "operationId": "list_boards",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Board"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "kanban"
        ],
        "operationId": "create_board",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BoardInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Board"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/boards/{id}": {
      "put": {
        "tags": [
          "kanban"
        ],
        "operationId": "update_board",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BoardInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Board"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "kanban"
        ],
        "operationId": "delete_board",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        }
      }
    },
    "/api/contacts/{id}": {
      "put": {
        "tags": [
          "contexts"
        ],
        "operationId": "save_contact",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Contact"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Contact"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "contexts"
        ],
        "operationId": "delete_contact",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        }
      }
    },
    "/api/context-links/{id}": {
      "put": {
        "tags": [
          "contexts"
        ],
        "operationId": "save_links",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContextLinks"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContextLinks"
                }
              }
            }
          }
        }
      }
    },
    "/api/contexts/{id}": {
      "put": {
        "tags": [
          "contexts"
        ],
        "operationId": "save_context",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkContext"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkContext"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "contexts"
        ],
        "operationId": "delete_context",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        }
      }
    },
...
    "/api/task-fields/{id}": {
      "put": {
        "tags": [
          "kanban"
        ],
        "operationId": "update_field",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskField"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskField"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Board": {
        "type": "object",
        "required": [
          "id",
          "name",
          "fieldId"
        ],
        "properties": {
          "fieldId": {
            "type": "string",
            "description": "Must be \"status\"; custom properties belong to contexts."
          },
          "id": {
            "type": "string"
          },
          "manualOrder": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "name": {
            "type": "string"
          },
          "sortMode": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "BoardInput": {
        "type": "object",
        "required": [
          "id",
          "name",
          "fieldId"
        ],
        "properties": {
          "fieldId": {
            "type": "string",
            "description": "Must be \"status\"; custom properties belong to contexts."
          },
          "id": {
            "type": "string"
          },
          "manualOrder": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          },
          "name": {
            "type": "string"
          },
          "sortMode": {
            "type": [
              "string",
              "null"
            ],
            "description": "Missing sorting fields preserve existing settings for older clients."
          }
        },
        "additionalProperties": false
      },
      "Contact": {
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "properties": {
          "email": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "ContextField": {
        "type": "object",
        "required": [
          "id",
          "name",
          "kind"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/ContextFieldKind"
          },
          "name": {
            "type": "string"
          },
          "options": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "value": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "ContextFieldKind": {
        "type": "string",
        "enum": [
          "identifier",
          "contact",
          "choice",
          "text",
          "date",
          "link",
          "note"
        ]
      },
      "ContextLinks": {
        "type": "object",
        "description": "Each owner has one independently revised attachment record. Empty values mask inheritance;\nabsent keys inherit. Field ids are scoped to a context, so unrelated schemas never collide.",
        "required": [
          "id",
          "contextIds",
          "overrides"
        ],
        "properties": {
          "contextIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "id": {
            "type": "string"
          },
          "overrides": {
            "type": "object",
            "additionalProperties": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              },
              "propertyNames": {
                "type": "string"
              }
            },
            "propertyNames": {
              "type": "string"
            }
          }
        },
        "additionalProperties": false
      },
...
      "FieldOption": {
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "TaskField": {
        "type": "object",
        "description": "Shared task status configuration; the historical wire shape is retained for sync retries.",
        "required": [
          "id",
          "name",
          "kind",
          "options"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "Must be \"status\"."
          },
          "kind": {
            "type": "string",
            "description": "Must be \"choice\"."
          },
          "name": {
            "type": "string"
          },
          "options": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldOption"
            }
          }
        },
        "additionalProperties": false
      },
...
      "WorkContext": {
        "type": "object",
        "required": [
          "id",
          "name",
          "fields"
        ],
        "properties": {
          "fields": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ContextField"
            }
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "additionalProperties": false
      }
    }
  },
...
}